refactor(ue): extract router + IHaybaMCPHandler, migrate 11 commands to FHaybaMCPLegacyHandler with namespaced aliases
