+++ packages/hayba/Plugins/HaybaMCPToolkit/Source/HaybaMCPToolkit/Public/HaybaMCPCommandHandler.h
@@ -1,50 +1,28 @@
 #pragma once
-
 #include "CoreMinimal.h"
-#include "Json.h"
-#include "UObject/UObjectIterator.h"
+#include "Dom/JsonObject.h"
 
-class UPCGSettings;
+class IHaybaMCPHandler;
 
 class FHaybaMCPCommandHandler
 {
 public:
-	FHaybaMCPCommandHandler();
-	~FHaybaMCPCommandHandler();
+    FHaybaMCPCommandHandler();
+    ~FHaybaMCPCommandHandler();
 
-	/** Parse a JSON command string, dispatch, and return JSON response string. */
-	FString ProcessCommand(const FString& CommandJson);
+    /** Register a handler for its commands. Called at startup. */
+    void RegisterHandler(TSharedRef<IHaybaMCPHandler> Handler);
 
-	/** Build a success response: {id, ok:true, data:{...}} */
-	static FString MakeOkResponse(const FString& Id, const TSharedPtr<FJsonObject>& Data);
+    /** Parse incoming TCP JSON, auth, dispatch, journal, return response JSON. */
+    FString ProcessCommand(const FString& CommandJson);
 
-	/** Build an error response: {id, ok:false, error:"..."} */
-	static FString MakeErrorResponse(const FString& Id, const FString& ErrorMessage);
+    /** Returns all registered command names. */
+    TArray<FString> GetAllCommands() const;
+
+    static FString MakeOkResponse(const FString& Id, const TSharedPtr<FJsonObject>& Data);
+    static FString MakeErrorResponse(const FString& Id, const FString& ErrorMessage);
 
 private:
-	// Command method signature
-	typedef FString (FHaybaMCPCommandHandler::*FCommandFunc)(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-
-	// Dispatch table
-	TMap<FString, FCommandFunc> CommandMap;
-
-	// Command implementations
-	FString Cmd_Ping(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_ListNodeClasses(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_GetNodeDetails(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_ListPCGAssets(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_ExportGraph(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_CreateGraph(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_ValidateGraph(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_ExecuteGraph(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_WizardChat(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_ImportLandscape(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-	FString Cmd_ReadNodeOutput(const TSharedPtr<FJsonObject>& Params, const FString& Id);
-
-	// Graph validation helper
-	TArray<TSharedPtr<FJsonValue>> ValidateGraphJson(const TSharedPtr<FJsonObject>& Graph) const;
-
-	// Helpers for node class discovery
-	TArray<UClass*> FindPCGExNodeClasses() const;
-	void GetPinInfo(const UClass* SettingsClass, TArray<TSharedPtr<FJsonValue>>& OutInputs, TArray<TSharedPtr<FJsonValue>>& OutOutputs) const;
+    TMap<FString, TSharedRef<IHaybaMCPHandler>> CommandToHandler;
+    TArray<TSharedRef<IHaybaMCPHandler>> Handlers;
 };
